requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-background-tasks==1.2.5
django-bootstrap4==2.0.1
django-compat==1.0.15
django-crispy-forms==1.9.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
mysqlclient
djangorestframework
paypal-checkout-serversdk
django-environ==0.4.5
python-amazon-sp-api